Remove Cnos from commons
