+++ src/sqlancer/common/TableQueryGenerator.java
@@ -1,11 +1,6 @@
 package sqlancer.common;
 
-import java.io.BufferedReader;
-import java.io.FileReader;
-
 import sqlancer.GlobalState;
-import sqlancer.cnosdb.CnosDBGlobalState;
-import sqlancer.cnosdb.query.CnosDBOtherQuery;
 
 public abstract class TableQueryGenerator extends FileExecutor {
     protected int total;
@@ -33,21 +28,6 @@ public abstract class TableQueryGenerator extends FileExecutor {
             }
         }
         throw new AssertionError();
-    }
-
-    public void runQueryFromFileCnos(String file, CnosDBGlobalState globalState) {
-        try {
-            FileReader fr = new FileReader(file);
-            BufferedReader br = new BufferedReader(fr);
-            String cur = br.readLine();
-            while (cur != null) {
-                globalState.executeStatement(new CnosDBOtherQuery(cur, null));
-                cur = br.readLine();
-            }
-            br.close();
-            fr.close();
-        } catch (Exception e) {
-        }
     }
 
     protected boolean isFinished() {
